Bug fix fot connectTo(address) in Terminal
If the address is not associated to a Site, create a new Site object for it.
Default encoding of Site is set to BBS::EncodingUnknown instead of Big5 now.
When the encoding is unknown, use Unicode (i. e. do ont use the tables to
convert, but use the two-byte code directly).