Fix double color drawing glich
Use QPixmap::fill() to fill the pixmap with the bg color.
QPainter::drawPixmap() and QPainter::drawPixmapFragments() both cause gliches
when the background of the top pixmap is transparent. Maybe a flag can do the
job, but I didn't find any.

Will this cause problems with real colors are introduced? We'll see.