Fix window sizing and displaying bug
Window is now set to fixed size after the tab is inserted.
The resizing when we change cell size will be a labor...

The size of the View is now promised to be larger than the backImage.
The backImage is drawn 1:1 in paintEvent().
This also solves the problem where the cursor, selection and URL line are
not in the right place (because the cells are promised to be of the same size,
now that the backImage is drawn 1:1).