Fix paste color
Paste color control sequence should match TARGET instead of SOURCE setting.

Therefore, we can't convert the data to string when we copy it, but need to
save to data directly, and convert it when the data are pasted.

Saving data is annoying, though. Implementing stream operator for BBS::Cell
is just too much work, so I just convert the needed settings into bytes,
and fit them directly into the QByteArray to save in QMimeData.