Fix #25: Paste function is broken
Introduced in commit ad1707088.

The text-insertion uses a QQueue instead of a plain QByteArray to
store the pending characters when between-character delay is enabled.
In such cases, therefore, the `bytes` variable will be empty (because
the `insertBuffer` instance variable is used instead).

The previous change naively checks for emptiness of `bytes` before
commting the insertion. While this does prevent crashing if `bytes`
is empty, it breaks situations in which `bytes` is empty, but
`inserBuffer` is not. The new implementation takes care of both
situations.

I'm feeling that having two kinds of insersion container is kind of
messy. This may need some refactoring in the future.