Implement screen refresh on font/color change
Fixes #13

The reason that the screen cannot be updated immediately is that
the actual settings commit occurs AFTER Controller::updateAll.

I think this is caused by a racing condition in QSettings and
QxtConfigWidget's implementation: When a setting is changed, a
`sync` is QUEUED into the event loop (instead of called directly,
to avoid redundant saving from occuring), but QxtConfigWidget's
`accept` calls inner widgets' `accept` DIRECTLY. This causes our
redraw event occurs before the settings actually change, resulting
the bug.

By using Qt::QueuedConnection to trigger the redraw instead, we
can queue the redraw event after the actual saving occurs, so that
the redraw can actually get the correct values.