Implement emoticon viewer
Add a list model class that is based on a JSON file. We'll
reuse that for the site manager later. A function is also
added in Globals (UJ::absoluteDataStoragePath) as a
convenience shorthand to obtain a writable path for data
files.

This one requires much more work than I anticipated, to
be honest.