Fix a bug where the MainWindow cannot size properly on OS X
For some reason, Qt under OS X ALWAYS reports that the window changes
from maximized to normal state on the inintial resizing. This makes
the window on Macs have a faulty initial window size on launch.

Since this is not an issue on any other platforms, and the behavior
doesn't really make enough case to justify time looking for a general
solution, I'll just make this a special case. Since windows on OS X
doesn't have a "maximized state" anyway, we'll just ignore this
particular event on Macs.