Initial working SSH based on QxtSshClient
An alternative standalone SSH implementation based solely on LibQxt
(with a bundled version of static-linked libssh2), OpenSSL, and zlib.

QxtSshClient is NOT included in the last release (0.6.2) of LibQxt, so
this branch is not buildable without compling LibQxt from source and
linking Qelly against it. We have two options now:

1. Bundle the tip version of LibQxt (maybe as a submodule) with Qelly,
   and static-link against it as a dependency. This eliminate the need
   of depending on the system's LibQxt but induce a lot of compiling
   overhead since we need to compile the whole LibQxt again, instead of
   being able to link against a pre-compiled binary.

2. Try to extract only SSH-related classes from LibQxt, and redistribute
   them only. This might be tedious since LibQxt uses a patched version
   of libssh2, and we don't know exactly what will happen if we don't
   link against it. But this approach has the advantage of smaller
   binary (at least on Linux) and only trivial overhead during building.

Licensing shouldn't be an issue in either case AFAIK since we use GPL
and LibQxt has LGPL.

TODO:
* Config backend and GUI frontend update.
* known_hosts implementation(?).

Related: #47 #30